requirements.txt
openai
langchain
streamlit
streamlit-extras
python-dotenv
huggingface-hub
pypdf